TSDB data import tool
Created a tool to import data formatted according to the Prometheus exposition format. The tool can be accessed via the TSDB CLI.

Addresses prometheus/prometheus#535

Rework tsdb data importer to be a two-pass system.
First pass assigns each sample a block index, and second pass
actually creates and writes the samples to the blocks.
This approach is required to deal with unsorted input data.

Signed-off-by: Dipack P Panjabi <dipack.panjabi@gmail.com>

Import correct label package

Signed-off-by: Dipack P Panjabi <dipack.panjabi@gmail.com>

Compare to check if samples are same or not, when testing blocks

Signed-off-by: Dipack P Panjabi <dipack.panjabi@gmail.com>

Close all readers

Signed-off-by: Dipack P Panjabi <dipack.panjabi@gmail.com>